refactor: organizing launch folder
