+++ solutions/devsprint-higor-silva-1/app/src/main/java/com/devpass/spaceapp/presentation/launch/launchpad/LaunchpadFragment.kt
@@ -1,0 +1,4 @@
+package com.devpass.spaceapp.presentation.launch.launchpad
+
+class LaunchpadFragment {
+}
